LayoutTests/http/tests/referrer-policy-script/no-referrer-when-downgrade/cross-origin-http.https-expected.txt

Tests the behavior of no-referrer-when-downgrade referrer policy when cross origin from HTTPS to HTTP.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is ""
PASS successfullyParsed is true

TEST COMPLETE
